refactor, test: 목적의 맞게 UserTest 테스트 클래스명 변경
